requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
